+++ app.js
@@ -358,6 +358,7 @@
           const { errhandler, handler, import_vcjs } = utils;
           const { fs, path } = sys;
           const { join } = path;
+          const { type: enginetype } = cosetting.general.engine;
 
           let output = handler.dataformat;
           try {
@@ -365,11 +366,13 @@
             let arr_process = [],
               arr_name = [];
             for (let val of arr_modname) {
-              let modpath = join(pathname, val);
-              if (fs.readdirSync(modpath).length > 0) {
-                let module = new cengine([modpath, val, curdir], obj);
-                arr_name.push(val);
-                arr_process.push(module);
+              if (val.indexOf(`${enginetype}_`) == 0) {
+                let modpath = join(pathname, val);
+                if (fs.readdirSync(modpath).length > 0) {
+                  let module = new cengine([modpath, val, curdir], obj);
+                  arr_name.push(val);
+                  arr_process.push(module);
+                }
               }
             }
             let arrrtn = await Promise.all(arr_process);