+++ engine/deskelectronjs/src/reaction.js
@@ -109,7 +109,7 @@ module.exports = async (...args) => {
             // let iscss = handler.check_empty(options.css);
 
             if (!isredirect) {
-              res.redirect(redirect);
+              res.redirect(redirect, sess);
               resolve(rtn);
             } else if (!isjson) {
               res.status(status).json(json);